assignment5(2)/requirements.txt
torch
transformers
Pillow
gradio